Fix several errors in tests run by Travis (#380)
* Declare file encoding to support special characters

* fix missing utf_8_encoder error in Travis tests

* Py 2.7 backwards-compat iterator

* ensure integer argument to torch.nn.functional.pad

* cast match.ceil result as integer

+++ torchaudio/functional.py
@@ -1,3 +1,5 @@
+# -*- coding: utf-8 -*-
+
 from __future__ import absolute_import, division, print_function, unicode_literals
 
 import math
@@ -1093,12 +1095,12 @@ def _compute_nccf(waveform, sample_rate, frame_time, freq_low):
     EPSILON = 10 ** (-9)
 
     # Number of lags to check
-    lags = math.ceil(sample_rate / freq_low)
+    lags = int(math.ceil(sample_rate / freq_low))
 
     frame_size = int(math.ceil(sample_rate * frame_time))
 
     waveform_length = waveform.size()[-1]
-    num_of_frames = math.ceil(waveform_length / frame_size)
+    num_of_frames = int(math.ceil(waveform_length / frame_size))
 
     p = lags + num_of_frames * frame_size - waveform_length
     waveform = torch.nn.functional.pad(waveform, (0, p))
@@ -1147,7 +1149,7 @@ def _find_max_per_frame(nccf, sample_rate, freq_high):
     to the first half of lags, then the latter is taken.
     """
 
-    lag_min = math.ceil(sample_rate / freq_high)
+    lag_min = int(math.ceil(sample_rate / freq_high))
 
     # Find near enough max that is smallest
